test: migrate common logic and compose UI tests to commonTest
- Move pure logic tests and compose tests to `commonTest` for KMP
- Replace JUnit dependencies with `kotlin.test`
- Replace Robolectric runners with Compose Multiplatform `runComposeUiTest`
- Add `compose.desktop.currentOs` to `jvmTest` to avoid Skiko library load errors
