+++ .idea/artifacts/richtext_editor.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="richtext-editor">
+    <output-path>$PROJECT_DIR$/richtext-editor/build/libs</output-path>
+    <root id="archive" name="richtext-editor.jar">
+      <element id="module-output" name="Arranger.richtext-editor.androidMain" />
+    </root>
+  </artifact>
+</component>
