+++ .idea/artifacts/richtext_editor_material3.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="richtext-editor-material3">
+    <output-path>$PROJECT_DIR$/richtext-editor-material3/build/libs</output-path>
+    <root id="archive" name="richtext-editor-material3.jar">
+      <element id="module-output" name="Arranger.richtext-editor-material3.androidMain" />
+    </root>
+  </artifact>
+</component>
