--- richtext-editor-material3/src/androidHostTest/resources/robolectric.properties
@@ -1,2 +1,0 @@
-sdk=36
-graphicsMode=NATIVE
