+++ richtext-editor/src/commonTest/kotlin/dev/mkeeda/arranger/richtext/editor/RichTextUndoManagerTest.kt
@@ -5,7 +5,7 @@ import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.booleans.shouldBeTrue
 import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.shouldBe
-import org.junit.Test
+import kotlin.test.Test
 
 class RichTextUndoManagerTest {
     private fun createSnapshot(text: String): EditorSnapshot {
