+++ richtext/src/commonTest/kotlin/dev/mkeeda/arranger/richtext/CharSequenceExtensionsTest.kt
@@ -3,7 +3,7 @@ package dev.mkeeda.arranger.richtext
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.shouldBe
-import org.junit.Test
+import kotlin.test.Test
 
 class CharSequenceExtensionsTest {
     @Test
