+++ richtext/src/commonTest/kotlin/dev/mkeeda/arranger/richtext/RichStringBatchEditTest.kt
@@ -2,7 +2,7 @@ package dev.mkeeda.arranger.richtext
 
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.shouldBe
-import org.junit.Test
+import kotlin.test.Test
 
 class RichStringBatchEditTest {
     @Test
